+++ src/models/claim_referral_code_dto_data.rs
@@ -20,7 +20,7 @@ pub struct ClaimReferralCodeDtoData {
     pub intent: models::IntentEnum,
     /// Message signedAt current timestamp (seconds since Unix Epoch)
     #[serde(rename = "signedAt")]
-    pub signed_at: f64,
+    pub signed_at: i64,
     /// The referral code to claim
     #[serde(rename = "code")]
     pub code: String,
@@ -30,7 +30,7 @@ impl ClaimReferralCodeDtoData {
     pub fn new(
         sender: String,
         intent: models::IntentEnum,
-        signed_at: f64,
+        signed_at: i64,
         code: String,
     ) -> ClaimReferralCodeDtoData {
         ClaimReferralCodeDtoData {
