+++ src/models/initiate_withdraw_dto_data.rs
@@ -35,7 +35,7 @@ pub struct InitiateWithdrawDtoData {
     pub nonce: String,
     /// Message signedAt current timestamp (seconds since Unix Epoch)
     #[serde(rename = "signedAt")]
-    pub signed_at: f64,
+    pub signed_at: i64,
 }
 
 impl InitiateWithdrawDtoData {
@@ -47,7 +47,7 @@ impl InitiateWithdrawDtoData {
         lz_destination_address: String,
         lz_destination_eid: models::LzDestinationEidEnum,
         nonce: String,
-        signed_at: f64,
+        signed_at: i64,
     ) -> InitiateWithdrawDtoData {
         InitiateWithdrawDtoData {
             account,
